simplify exported functions for tests
